Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 
 int main(){
-    float r;
-    scanf("%f",&r);
+    int r;
+    scanf("%d",&r);
     printf("Area: %.2f";3.14*r*r);
     return 0;
 }
